Added drop-down selection for values on the review page.
Combined two url patterns in urls.py.

Moved icons for editing and deleting cart items to the left of the page in cart_items_table.html.
(This was so they are still aligned if different items have different amounts of data to display.

Fixed review_form.html to send the form to the correct url. ('./' instead of '../')

Removed commented microdissected code from review_form.html.


git-svn-id: svn+ssh://gleek.ecs.baylor.edu/matrr/trunk@32 8ba6b188-5121-491e-b013-2137287ea23c