Fixed broken links in the footers.  Terrible Django-fu.
The issue was that f.open('path/to/static.txt', r) wasn't working with the inconsistent directory structures between the dev and production environments.  path/to was pointing to static/, which worked as long as the working directory was the project directory, which didn't seem to be the case on gleek. 

I tried modifying the settings file to point to different static dir's given the current environment, but ran into issues with the target static directory being one of the STATICFILES_DIRS search targets.  At that point I would have had to create a static folder on each dev environment.  For 3 files/pages, it simply wasn't worth the hassle of finding an eligent solution.

git-svn-id: svn+ssh://gleek.ecs.baylor.edu/matrr/trunk@145 8ba6b188-5121-491e-b013-2137287ea23c