plotting.boxplot_m2de() now creates a boxplot for etoh_intake, veh_intake, total_pellets and weights, creates a thumbnail for each and dumps them into subdirectories within static/images
cohort_image filter got more complex.  It now accepts a parameter, 'etoh' 'veh' 'pellet' 'weight' and 'all'.  it will check if the full image for the given parameter exists  (!!not the thumbnail.  if the thumbnail doesnt exist but the image does, the thumbnail wil 404!!).  If it the full image exists it returns a dictionary indexed by 'etoh' et al. and valued with the graph title.  These test keys title values are a variable in the filter which templates use.

changed chorts template to use the thumbnail.  

Cohorts displays a title from basic_filters.cohort_image and each thumbnail, which links to the full image.

the thumbnail div container limits the width to 3 thumbnails per row.  Smaller thumbnails will allow more per row



git-svn-id: svn+ssh://gleek.ecs.baylor.edu/matrr/trunk@159 8ba6b188-5121-491e-b013-2137287ea23c