Remove some "TODO:" comments out of some static files that I'm not maintaining.
Moved several templates/matrr/tools/*.html files into an appropriate subdirectory, based on which tool the template is being used for.

Added the models, views, urls, forms and static files to support the hormone tools pages.  This does not include the methods to create the MonkeyHormoneImages or CohortHormoneImages.

None of this code has been tested yet.