Finally added matrr.utils.stackplot.  It's a pretty spiffy object/method/script.
It's kind of a jerk though.  I couldn't figure out how to make a Figure that worked independent of the environment.  matplotlib.pyplot.figure creates 1 figure in the world, so anytime I tried to make more than 1 stackplot it just rendered them on top of eachother on the same figure.  This is a known (to me) issue of using almost anything pyplot.  Anytime I tried to use matplotlib's Figure object I got errors about not having... something.  I forget what, but I never figured out how to fix it.

SO, I made a __main__() call, so I can run my function from the command line and have it dump images on execution.  A hack, for true, but if it works that's fine with me.

That said, the __main__ bit isn't actually tested, but the object works and the function should work fine.

-jf